requirements.txt
langchain
langchain-ollama
langchain-core
langchain-community
python-dotenv
langserve
sse-starlette
bs4
PyPDF
faiss-cpu
chromadb
langchainhub
transformers
tensorflow
torch 
numpy
pandas
scikit-learn
matplotlib
seaborn
FastAPI
uvicorn
datasets
langchain-openai